% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchScanning.R
\name{batchScanning}
\alias{batchScanning}
\title{Scanning images in batch}
\usage{
batchScanning(
  path,
  fflag = "controle",
  resdir = "resdir",
  bandwidth,
  fpicks,
  lengthLimit = 30,
  gridscan = 10
)
}
\arguments{
\item{path}{The file directory containing the images for analysis}

\item{fflag}{The pattern observed in the name of the files}

\item{resdir}{The name of the directory that will be created for the storage
of the resulting files}

\item{bandwidth}{describe}

\item{gridscan}{An integer value to define the size of the grid that will be
applicated to the segmented image.}

\item{fp}{describe}

\item{lenghLimit}{An integer value to define the minimum size to be consider
as a neurite length. The value is in number of pixels.}
}
\value{
A file directory containing the resulting segmented images along with
a .csv file including the quantitative results obtained by the function
neuriteScan.
}
\description{
Function that scans images in batch mode.
}
\examples{

}
\seealso{
\code{\link{channelSegmentation}}

\code{\link{neuriteScan}}
}
