% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelSegmentation.R
\name{channelSegmentation}
\alias{channelSegmentation}
\title{Two-channel segmentation of an image object}
\usage{
channelSegmentation(
  img,
  nwin = 3,
  smooth = 3,
  minShape = 40,
  maxShape = 800,
  channelnuc = "blue",
  channelcyto = "green",
  bandwidth = 20,
  fpicks = 0.01,
  colnuc = "cyan",
  colcyto = "white",
  colscheme = 3
)
}
\arguments{
\item{img}{An image object as input to be segmented.}

\item{nwin}{describe}

\item{smooth}{describe}

\item{minShape}{describe}

\item{maxShape}{describe}

\item{channelnuc}{A character variable to define the channel color for
the nucleus staining.}

\item{channelcyto}{A character variable to define the channel color for
the cytoplasm staining.}

\item{bandwidth}{describe}

\item{fpicks}{describe}

\item{colnuc}{A character variable to define the color of the nucleus in the
resulting painted image.}

\item{colcyto}{A character variable to define the color of the cytoplasm in
the resulting painted image.}

\item{colscheme}{An integer to define the scheme of colors in the resulting
painted image.}
}
\value{
A list containing the original image, the resulting grayscale images,
the binary masks and the segmented final image.
}
\description{
Function that runs two-channel segmentation for one image.
}
\examples{
image <- system.file("images", "cells.jpg", package = "neuriteScan")
img <- EBImage::readImage(image)
segImage <- channelSegmentation(img=img, bandwidth=10, fpicks = 0.01)
EBImage::display(segImage$paintedImage)

}
\seealso{
\code{\link[EBImage{channel}]}

\code{\link[EBImage{paintObjects}]}

\code{\link{preProcessing}}
}
