% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridLines.R
\name{getGridLines}
\alias{getGridLines}
\title{Returning some grids in the resulting image}
\usage{
getGridLines(mask, gridscan = 10, thickness = 1)
}
\arguments{
\item{mask}{A binary image used as mask, such as mask contained in the
resulting list of the channelSegmentation function.}

\item{gridscan}{An integer value to define the size of the grid that will be
applicated to the segmented image.}

\item{thickness}{An integer value to define the thickness of the grid lines.}
}
\value{
An binary image containing grid lines.
}
\description{
Function that performs displaying grid lines in an image object.
}
\examples{
grid <- getGridLines(segImage$cytomask)
paintedImage <- segImage$paintedImage
paintedImage[grid==1] <- 1
EBImage::display(paintedImage)

}
\seealso{
\code{\link{channelSegmentation}}

\code{\link{neuriteScan}}
}
