% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuriteScan.R
\name{neuriteScan}
\alias{neuriteScan}
\title{Scan objects in a segmented image}
\usage{
neuriteScan(segImage, gridscan = 10, lengthLimit = 30)
}
\arguments{
\item{segImage}{A segmented image obtained using the channelSegmentation
function. That image will be used as input for neurite scanning.}

\item{gridscan}{An integer value to define the size of the grid that will be
applicated to the segmented image.}

\item{lengthLimit}{An integer value to define the maximum size to be consider
as a neurite length. The value is in number of pixels.}
}
\value{
A list containing the final results. This list contains
the results for the total count of neurites, the total count of cells, the
number of neurites per cytoplasm area and the number of neurites per cell.
}
\description{
Function that performs neurite scanning in a segmented image.
}
\examples{
results <- neuriteScan(segImage,gridscan=10, lengthLimit = 30)

}
\seealso{
\code{\link{channelSegmentation}}
}
