% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preProcessing}
\alias{preProcessing}
\title{Image Pre-Processing}
\usage{
preProcessing(img)
}
\arguments{
\item{img}{An image object as input to be pre-processed.}
}
\value{
An improved image with adjusted contrast and less noise.
}
\description{
Function that performs image pre-processing with the aim of minimize the
noise and contrast problems that might be associated with a poor staining or
bad image quality.
}
\details{
This function uses skimage and numpy functions. Skimage and numpy are
modules imported from python via reticulate package.
}
\examples{
image <- system.file("images", "cells.jpg", package = "neuriteScan")
img <- EBImage::readImage(image)
img <- preProcessing(img)
EBImage::display(img)

}
\seealso{
\code{\link[EBImage{clahe}]}
}
