% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRGB.R
\name{readRgbChannels}
\alias{readRgbChannels}
\title{Merge image files into RGB}
\usage{
readRgbChannels(red = NULL, green = NULL, blue = NULL)
}
\arguments{
\item{red}{A path to the file containing an image for the red channel.}

\item{green}{A path to the file containing an image for the green channel.}

\item{blue}{A path to the file containing an image for the blue channel.}
}
\value{
The image in RGB format.
}
\description{
Function that performs files merging to RGB format.
}
\examples{
blue <- system.file("images", "nucleus.jpg", package = "neuriteScan")
green <- system.file("images", "cytoplasm.jgp", package = "neuriteScan")
img <- readRgbChannels(green = green, blue = blue)
EBImage::display(img)

}
\seealso{
\code{\link{channelSegmentation}}

\code{\link[EBImage{channel}]}

\code{\link[EBImage{readImage}]}

\code{\link[EBImage{rgbImage}]}
}
